+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DE-0L4kj.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-Dbgk4ZAA.js";import{_ as a}from"./mermaid.core-Bp44dlkC.js";import"./chunk-5VM5RSS4-BxDpa26T.js";import"./chunk-XXDRQBXY-tEvZAHbI.js";import"./chunk-POPQ4Y6H-DjsNS5Y7.js";import"./chunk-F27PBJKO-B0okP9oH.js";import"./channel-BzbrrejQ.js";import"./index-yqlljjQO.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};